pyproject.toml
[project]
name = "etso"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.16",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-openai>=0.3.29",
    "langgraph>=0.6.4",
    "numpy>=2.3.2",
    "openai>=1.99.6",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "structlog>=25.4.0",
    "tabulate>=0.9.0",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
]
